+++ docs/changelog.md
@@ -7,6 +7,7 @@
 - [alerts/annotations](checks/alerts/annotation.md) and [rule/label](checks/rule/label.md) will
   now report when a label or annotation is required but set to an empty value.
 - Fixed handling of `# pint ignore/line` comments on line including multiple `#` characters.
+- When reporting problems all messages will now use `publicURI` from each `prometheus` definition.
 
 ## v0.59.0